% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_load.R
\name{amphi_list_files}
\alias{amphi_list_files}
\title{Llista arxius de diferents carpetes de dades.}
\usage{
amphi_list_files(
  home_folder,
  recursive = TRUE,
  pattern = c("xlsx", "xls", "csv")
)
}
\arguments{
\item{home_folder}{Objecte R amb la ruta a les carpetes de dades.}

\item{recursive}{Lògic (TRUE per defecte). Llista els arxius de forma recursiva.}

\item{pattern}{Carrega un format d'extensió específic entre 'xlsx', 'xls' i 'csv'.}
}
\value{
Una llista d'arxius estructurada per 'carpetes' anomenades.
}
\description{
Llista arxius de diferents carpetes de dades donat un 'home_folder'.
}
