% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{amphi_setup_dirs}
\alias{amphi_setup_dirs}
\title{Configura els directoris i carrega els paquets del projecte}
\usage{
amphi_setup_dirs(
  base_path = getwd(),
  dirs_to_create = c("data", "R", "output", "capitols")
)
}
\arguments{
\item{base_path}{La ruta base del projecte. Per defecte, el directori de treball actual.}

\item{dirs_to_create}{Un vector de caràcters amb els noms dels directoris a crear.}
}
\value{
No retorna cap valor, però mostra un missatge per cada directori creat.
}
\description{
Aquest script s'encarrega de la configuració inicial del projecte,
incloent la creació de directoris i la càrrega de paquets necessaris.

Aquesta funció comprova si existeixen els directoris clau del projecte i els crea si no existeixen.
}
\examples{
# Executar amb els directoris per defecte
# amphi_setup_dirs()
}
