% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{amphidata_sticker}
\alias{amphidata_sticker}
\title{Genera un Hex Sticker per al Paquet 'amphidata'}
\usage{
amphidata_sticker(
  x_amphi_pos = -0.15,
  y_amphi_pos = 0.1,
  x_data_pos = 0.45,
  y_data_pos = -0.1,
  font1_family = "LM Roman",
  font2_family = "data_font",
  color1 = "#8E1F2F",
  color2 = "#F0BC42",
  size1_geom = 35,
  size2_geom = 28,
  output_format = "png",
  output_width = 4,
  output_height = 4,
  filename = "amphi_logo"
)
}
\arguments{
\item{x_amphi_pos}{Coordenada X (entre -1 i 1) per a 'amphi'. Valor per defecte: -0.15.}

\item{y_amphi_pos}{Coordenada Y (entre -1 i 1) per a 'amphi'. Valor per defecte: 0.1.}

\item{x_data_pos}{Coordenada X (entre -1 i 1) per a 'data'. Valor per defecte: 0.45.}

\item{y_data_pos}{Coordenada Y (entre -1 i 1) per a 'data'. Valor per defecte: -0.1.}

\item{font1_family}{Família de la font per a 'amphi'. Valor per defecte: 'LM Roman'.}

\item{font2_family}{Família de la font per a 'data'. Valor per defecte: 'data_font'.}

\item{color1}{Color principal del fons de l'hexàgon. Valor per defecte: \verb{#8E1F2F}.}

\item{color2}{Color secundari del text i la vora. Valor per defecte: \verb{#F0BC42}.}

\item{size1_geom}{Mida (geom size) per a la paraula 'amphi'. Valor per defecte: 35.}

\item{size2_geom}{Mida (geom size) per a la paraula 'data'. Valor per defecte: 28.}

\item{output_format}{Format de sortida: 'svg' o 'png'. Valor per defecte: 'png'.}

\item{output_width}{Amplada del fitxer de sortida (en polzades). Valor per defecte: 4.}

\item{output_height}{Alçada del fitxer de sortida (en polzades). Valor per defecte: 4.}

\item{filename}{Nom del fitxer de sortida (sense extensió). Valor per defecte: "hexsticker_amphidata_final".}
}
\value{
Desa un fitxer a la carpeta de treball i retorna de manera invisible el nom del fitxer creat.
}
\description{
Crea un 'hex sticker' personalitzat per al paquet 'amphidata', combinant les paraules 'amphi' i 'data' amb un disseny i colors específics.
}
\details{
La funció utilitza \code{ggplot2} per crear el disseny intern de l'hexàgon, que inclou les paraules 'amphi' i 'data' amb fonts personalitzades ('LM Roman' i 'Dancing Script', que es carreguen automàticament des de Google Fonts si no estan disponibles localment). Posteriorment, el paquet \code{hexSticker} s'encarrega de renderitzar el gràfic en el format hexagonal final. La funció permet una àmplia personalització dels elements visuals com posicions, colors, mides de font i dimensions de sortida.
}
\examples{
\dontrun{
# Genera la imatge final de 4x4 polzades (amb els valors per defecte)
crear_amphidata_sticker() 
}
}
\seealso{
\code{\link[hexSticker]{sticker}}, \code{\link[ggplot2]{ggplot}}, \code{\link[showtext]{font_add_google}}
}
