% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_imputacio_missforest.R
\name{imputacio_missforest}
\alias{imputacio_missforest}
\title{Imputació de dades amb \code{missForest}}
\usage{
imputacio_missforest(
  df,
  seleccio_variables,
  grup_by,
  grup_by_reserva,
  optim_mtry = FALSE,
  ntree = 100,
  maxiter = 10,
  verbose = FALSE,
  set_seed = 19810424,
  report_imputacio = TRUE,
  retornar_original = FALSE
)
}
\arguments{
\item{df}{Un data frame.}

\item{seleccio_variables}{Columnes a imputar (tidy-select).}

\item{grup_by}{Columnes per agrupar la imputació (tidy-select).}

\item{grup_by_reserva}{Columnes per agrupar si queden NAs (tidy-select).}

\item{optim_mtry}{Booleà (\code{TRUE}/\code{FALSE}) per optimitzar \code{mtry}.}

\item{ntree}{Nombre d'arbres a créixer al bosc.}

\item{maxiter}{Nombre màxim d'iteracions de l'algorisme.}

\item{verbose}{Lògic. Si és \code{TRUE}, mostra missatges de \code{missForest}.}

\item{set_seed}{Un número per fixar la llavor aleatòria i assegurar la reproductibilitat.}

\item{report_imputacio}{Retorna un report de la imputació? (default: TRUE).}

\item{retornar_original}{Retorna el data frame original juntament amb l'imputat? (default: FALSE).}
}
\value{
Una llista que conté:
\itemize{
\item \code{imputed_df}: El data frame amb els valors imputats i les columnes 'flag_*'.
\item \code{original_df}: (Opcional) El data frame original.
\item \code{report}: (Opcional) Una llista amb un resum i el detall de la imputació.
}
}
\description{
Aquesta funció realitza una imputació de valors perduts utilitzant l'algorisme
\code{missForest}. Permet l'optimització de paràmetres i l'agrupació de dades.
Genera columnes de "flags" per a cada variable imputada.
}
\examples{
# Crear un data frame d'exemple amb NAs
df_exemple_miss <- data.frame(
  grup = as.factor(c("A", "A", "B", "B", "A", "B")),
  valor1 = c(1, NA, 3, 4, 5, NA),
  valor2 = c(NA, 2, 3, NA, 5, 6)
)

# Imputar amb missForest sense agrupar
resultat_miss <- imputacio_missforest(
  df = df_exemple_miss,
  seleccio_variables = c(valor1, valor2)
)

# Imputar amb missForest agrupant per 'grup'
resultat_miss_agrupat <- imputacio_missforest(
  df = df_exemple_miss,
  seleccio_variables = c(valor1, valor2),
  grup_by = grup
)
}
