% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tab_summary.R
\name{tab_summary}
\alias{tab_summary}
\title{Calcula estadístiques descriptives per a un \code{data.frame}.}
\usage{
tab_summary(
  df,
  grup_by = NULL,
  addicio_info = NULL,
  seleccio_variables,
  stats_adicionals = FALSE,
  sd_num = NULL,
  q_lower = NULL,
  q_upper = NULL,
  na.rm = TRUE,
  bind_rows = TRUE,
  digits = 2
)
}
\arguments{
\item{df}{El \code{data.frame} o \code{tibble} d'entrada.}

\item{grup_by}{Les variables per agrupar les dades. També pot utilitzar la sintaxi de selecció de \code{dplyr}. Si es deixa com a \code{NULL}, les estadístiques es calculen per a tot el \code{data.frame}.}

\item{addicio_info}{Columnes addicionals del dataframe original per incloure a la taula de resultats. Útil per afegir informació contextual que és constant dins de cada grup.}

\item{seleccio_variables}{Les variables numèriques a analitzar. Es pot utilitzar la sintaxi de selecció de \code{dplyr} (per exemple, \code{starts_with("var")}, \code{contains("grup")}).}

\item{stats_adicionals}{Lògic. Si és \code{TRUE}, calcula estadístiques addicionals com curtosi i asimetria.}

\item{sd_num}{Un valor numèric per definir el nombre de desviacions estàndard per a les estadístiques de límit inferior i superior.}

\item{q_lower}{Un valor numèric per a calcular un quantil de límit inferior.}

\item{q_upper}{Un valor numèric per a calcular un quantil de límit superior.}

\item{na.rm}{Un valor lògic. Si és \code{TRUE} (per defecte), els valors \code{NA} s'eliminen abans de tots els càlculs.}

\item{bind_rows}{Un valor lògic. Si és \code{TRUE} (per defecte), la llista de \code{tibbles} es combina en un sol \code{tibble}, afegint una columna anomenada \code{variable} amb el nom de la variable analitzada.}

\item{digits}{El nombre de dígits decimals a mostrar per a les variables \code{double}. Per defecte és 2. Les variables \code{integer} no s'arrodoneixen.}
}
\value{
Un \code{tibble} que conté les estadístiques descriptives calculades per a cada variable i grup.
}
\description{
Calcula un conjunt complet d'estadístiques descriptives per a variables numèriques.
La funció està optimitzada per al seu ús amb el \code{tidyverse}.
}
\details{
Aquesta funció utilitza \code{dplyr::summarise()} i \code{lapply()} per aplicar de manera eficient
múltiples funcions a diverses variables. Les proves d'asimetria i curtosi només s'executaran
si s'especifiquen a \code{stats_adicionals}.
}
\examples{
\dontrun{
# Carreguem les dades amb una funció del paquet
df_golvin <- load_dimensions_golvin()

# Calcula estadístiques descriptives per grup
tab_summary(
  df = df_golvin,
  seleccio_variables = starts_with('amplada'),
  grup_by = pais,
  stats_adicionals = TRUE,
  sd_num = 3,
  q_lower = .05,
  q_upper = .95,
  na.rm = TRUE,
  bind_rows = TRUE,
  digits = 2)
}

}
\seealso{
\code{\link[e1071]{skewness}} i \code{\link[e1071]{kurtosis}} per al càlcul d'asimetria i curtosi.
}
